BNGP-527-CYA Taking care of review requests

+++ packages/webapp/src/routes/land/check-management-monitoring-details.js
@@ -31,7 +31,7 @@ function getFormattedDate (dateString) {
   const date = moment(dateString)
 
   return date.toDate().toLocaleDateString('en-GB', {
-    day: 'numeric', month: 'short', year: 'numeric'
+    day: 'numeric', month: 'long', year: 'numeric'
   })
 }
